corrections in requests - flats and bookings only of certain user, corrections in frontend etc.

+++ src/main/java/edu/pw/react/project/backend/model/FlatEntity.java
@@ -22,7 +22,7 @@ public class FlatEntity {
     private Long id;
 
     @Column(name = "owner_of_room")
-    private int owner_of_room;
+    private Long owner_of_room;
 
     @Column(name = "name_of_room")
     private String name_of_room;
